feat: filter mcp tools based on setting
